Refactoring de la classe Triangle, externalisation de l'instanciation du Program et transformation en une abstract classe. Création de la classe BasicTriangle qui est un Triangle à la position donné et utilisant un BasicProgram.
