+++ app/src/main/java/cafetcraftteam/rpgsoftware/gui/open_gl_es/Drawable.java
@@ -1,0 +1,9 @@
+package cafetcraftteam.rpgsoftware.gui.open_gl_es;
+
+/**
+ * Created by Gautier on 03/03/2016.
+ */
+public interface Drawable
+{
+    void draw(float[] mvpMatrix);
+}
