+++ app/src/main/java/cafetcraftteam/rpgsoftware/gui/open_gl_es/MyGLRenderer.java
@@ -48,7 +48,7 @@ public class MyGLRenderer implements GLSurfaceView.Renderer {
         // Set the background frame color
         GLES20.glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
 
-        mTriangle = new Triangle();
+        mTriangle = new BasicTriangle();
     }
 
     /**
